// src/enums/ability-keyword.ts
export enum AbilityKeyword {
	Animal = 'Animal',
	Area = 'Area',
	Attack = 'Attack',
	Charge = 'Charge',
	Defense = 'Defense',
	Earth = 'Earth',
	Fire = 'Fire',
	Force = 'Force',
	Green = 'Green',
	Magic = 'Magic',
	Melee = 'Melee',
	Persistent = 'Persistent',
	Psionic = 'Psionic',
	Ranged = 'Ranged',
	Resistance = 'Resistance',
	Void = 'Void',
	Weapon = 'Weapon'
}